Was static class in interface "Field", now "Property" interface+impl.
