+++ kie-dmn-feel/src/main/java/org/kie/dmn/feel/lang/CustomType.java
@@ -1,33 +1,9 @@
 package org.kie.dmn.feel.lang;
 
-import java.util.List;
+import java.util.Map;
 
 public interface CustomType extends Type {
 
-    List<Field> fields();
+    Map<String, Property> getProperties();
 
-    public static class Field {
-        private final String name;
-        private final Type type;
-        public Field(String name, Type type) {
-            super();
-            this.name = name;
-            this.type = type;
-        }
-        
-        public String getName() {
-            return name;
-        }
-        
-        public Type getType() {
-            return type;
-        }
-
-        @Override
-        public String toString() {
-            StringBuilder builder = new StringBuilder();
-            builder.append("Field [name=").append(name).append(", type=").append(type).append("]");
-            return builder.toString();
-        }
-    }
 }
