+++ kie-dmn-feel/src/main/java/org/kie/dmn/feel/lang/Property.java
@@ -1,0 +1,10 @@
+package org.kie.dmn.feel.lang;
+
+
+public interface Property {
+
+    String getName();
+
+    Type getType();
+
+}
