+++ kie-dmn-feel/src/main/java/org/kie/dmn/feel/lang/impl/MapBackedType.java
@@ -1,16 +1,18 @@
 package org.kie.dmn.feel.lang.impl;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import org.kie.dmn.feel.lang.CustomType;
+import org.kie.dmn.feel.lang.Property;
 import org.kie.dmn.feel.lang.Type;
 import org.kie.dmn.feel.parser.feel11.ParserHelper;
 
 public class MapBackedType implements CustomType {
     
-    private List<Field> fields = new ArrayList<>();
+    private Map<String, Property> properties = new HashMap<>();
     
     public MapBackedType() {
     }
@@ -21,12 +23,12 @@ public class MapBackedType implements CustomType {
      */
     public MapBackedType(Map<String, ?> map) {
         map.entrySet().stream()
-            .map( kv -> new Field( kv.getKey(), ParserHelper.determineTypeFromClass( kv.getValue().getClass()) ) )
-            .forEach( f -> fields.add(f) );
+            .map( kv -> new PropertyImpl( kv.getKey(), ParserHelper.determineTypeFromClass( kv.getValue().getClass()) ) )
+            .forEach( f -> properties.put( f.getName(), f ) );
     }
     
-    public MapBackedType(List<Field> fields) {
-        this.fields.addAll(fields);
+    public MapBackedType(List<Property> properties) {
+        properties.stream().forEach(p -> this.properties.put(p.getName(), p) );
     }
 
     @Override
@@ -44,13 +46,13 @@ public class MapBackedType implements CustomType {
         return null;
     }
     
-    @Override
-    public List<Field> fields() {
-        return fields;
-    }
-    
     public MapBackedType addField(String name, Type type) {
-        fields.add(new Field(name, type));
+        properties.put( name, new PropertyImpl(name, type) );
         return this;
+    }
+
+    @Override
+    public Map<String, Property> getProperties() {
+        return properties;
     }
 }
